cloud_storage: fix term last offset archive queries
Consider the example below:
        [--------]  [--------]
        ^           ^    ^
Term:   T1          T1   T2
Offset: A           B  X C
                       ^
                       |________ _archive_start_offset

Previously, when looking for the last offset in T1, we'd correctly pick
the second spillover manifest in the diagram and we would attempt to
create a cursor starting at B (i.e. the start of the selected spill
manifest). This request is not fulfilled as the cursor refuses to answer
queries below the archive start offset. The fix here, is to init the
cursor with the start offset of the archive iff that lies within the
same spillover manifest.