rptest: Ensure positive truncate offset in DR test
- In the truncate at segment boundary test, if the first segment is
randomly chosen as the segment to truncate at, subtractng one from its
base offset would mean to truncate at a negative offset which would
cause KCL to error as it cannot handle negative offsets as truncation
points.

- Fix is to never truncate at first segment when this at segment
boundary parameter is chosen.

- Additional validation is performed in KCL.py to raise a more obvious
exception in the case negative offsets are passed as parameters to the
delete_records method.

- Fixes: #11847