cluster: Fix for incorrect end_offset reported
- This change reverts a previous attempt to fix rm_stm reporting
invalid_lso (in change 20285dff0f35f07c9b30dbcb5a79907e9696f714), which
sets `booststrap_committed_offset` in apply_snapshot() to fix the
case where invalid_lso is indefinitely returned on a node that had
just restarted, applied a snapshot, but no data was produced onto it.

- The change was however incorrect, as the bootstrap committed offset is
expected to be the value of a complete read of the log, and the value of
the offset in the rm_stm snapshot at the time of reading it at startup,
does not necessarily reflect this.

- The solution is to at startup wait until the consensus later has
modified the committed offset.

(cherry picked from commit 88cfa564f14def6b16ebb1553fbeb9ae76f4350d)