kafka/connecton_context: Add a sharded_abort_source
The `sharded_abort_source` is subscribed to the
`net::server::abort_source`, so that it is aborted when the server
is aborted, but can also be manually aborted earlier.

Some requests such as fetch require results from several partitions,
which live on disparate cores; a sharded abort source enables that.

Signed-off-by: Ben Pope <ben@redpanda.com>