c/archival_stm: use replace manifest command to reset stm state
Previously a `reset_metadata` command was followed up with the commands
setting up the `archival_metadata_stm` state (i.e. partition_manifest).
Since now the manifest is binary serialized and kept small in size it is
possible to serialize it as a part of a command.

Leveraging recently introduced `replace_manifest` command to reset the
`archival_metadata_stm` state.

Signed-off-by: Michal Maslanka <michal@redpanda.com>