services/redpanda: raise_on_cloud_storage_inconsistencies
this method is similar to stop_and_scrub_cloud_storage, it uses
rp-storage-tool to check uploaded data. the main difference is that is
does not stop the cluster, and needs a list of anomalies to raise for.
It's meant to be used while a cluster is running, when the tool might
see some temporary inconsistencies