r/consensus: check for last applied inconsistency in raft
If `last_applied` offset persisted in kvstore is greater than the log
dirty offset it indicates the inconsistency. The inconsistency may be a
result of intentional removal of a log segments from Redpanda data
directory. In this scenario the `last_applied` offset must be removed
from kv-store to prevent it from updating the committed offset which may
result in not committed batches being applied to stm.

Signed-off-by: Michal Maslanka <michal@redpanda.com>