archival: Guarantee that the ntp_archiver uses...
the same archiver term for the segment name and metadata. Previously,
the ntp_archiver used cached current term id to create uploads. This
might potentially introduce a race condition when the leadership
transfer is prepared and the ntp archiver is paused but the same node is
becoming a leader again. In this case we may create a segment name with
one archiver term and the manifest metadata with the next term.