cluster: Eviction stm continue to evict if it can
- This fixes an annoyance observed in CI where the eviction stm will not
continue to evict in situations where it evicted up until a max
collectible offset which is less then the requested.

- The eviction fiber is woken up once per event so if it did not
truncate up until the desired eviction offset it would on the next
event.

- This however can be improved upon. It can identify situations in which
is was not sucessful in eviction up until the desired event and instead
of indefinitely waiting until the next wakeup, wait with a static
timeout. This will cause the eviction fiber to wakeup after some time
and retry to evict more data.

- Fixes: #11918