Assert on updates set in reconciliation
In members_backend::reconcile, we call _updates.front() which
requires that updates be non-empty. Though we did check that
condition earlier, we pass through a suspension point since then
(_raft0->linearizable_barrier()) so it is not obvious from local
reasoning that the _updates array will _always_ be empty and
that this access will be safe.

So we assert on it.