rpk profile create: require a name without --from-cloud
IME using these commands a lot in the past few weeks, creating a profile
with the name "default" is confusing -- I always rename the profile
immediately after creation. Rather than supporting this, we can just
error immediately without a name.

The one exception is --from-cloud, where we can get the name directly
from the remote cluster. If an existing profile name collides with the
cloud cluster name, we exit with an error indicating to specify a name.