tests: fixed querying for the node in decommission test
In current redpanda version node ids are dynamically assigned we can not
use a node id as an index in the `RedpandaService` nodes list as it may
result in incorrect results. In this particular example a node that was
stopped was not the one that was decommissioned, leading to some of the
rf=1 partitions being not available which prevented test from finishing.

Fixes: #11751

Signed-off-by: Michal Maslanka <michal@redpanda.com>
(cherry picked from commit 4ec279e66096d5f582a0d3f4343f49cc1a68ccad)