c/backend: wait for leader to be elected before finishing abort cancel
When forcibly aborting reconfiguration we should wait for the new leader
to be elected in the configuration that the partition was forced to.
This way we can be certain that the new configuration will finally be
replicated to the majority of nodes even tough the leader may not
exists at the time when configuration is replicated.

Fixes: #9243

Signed-off-by: Michal Maslanka <michal@redpanda.com>
(cherry picked from commit 3948312e2c2af16049b80987344c3822b4c5d0f2)