kafka: Disable use of separate fetch scheduling group
This partially reverts 9a93a9c22238e145dfa9d7fe297eb494d7c5f0bf

While the original motiviation isn't invalidated we have now found a
counter example where the extra fetch groups makes things worse overall.

`ManyPartitionTest` fails on ARM with the extra group but passes
without. With the group in use CPU util hits 100% and grinds everything
to halt.

Fetch seems to be a lot slower on ARM. Hence, with the guaranteed share
of the extra group the whole system gets affected and hits CPU limits.

Because this is incredibly hard to reason about and it wasn't the core
fetch optimization we decided to revert back to keeping it disabled by
default.

We still keep the option around as it might be useful potentially in
corner cases.

Fixes https://github.com/redpanda-data/redpanda/issues/10507