pb_planner: simplify iterator check
Gets rid of the comma operator in the loop that is expected to throw
when the iterator is invalidated. Instead adds an explicit check that
does the same.

ASAN traces in #11664 are ideally not possible because the
freeing stack clearly bumps the topic map revision right after erase
that deallocates the memory, so any subsequent operator the iterator
show throw right away.

The code is unnecessarily complex with the comma operator, so getting
rid of it in the hopes that there is some sort of miscompilation.