r/state_machine: share linearizable barrier result
In order to prevent contention implemented sharing linearizable barrier
result between contended callers. Instead of calling linearizable
barrier multiple times a caller will wait for the result of a barrier
that is already being executed. This doesn't change the current
semnatics of linearizable barrier as either way a caller must check the
returned offset if they want to wait for the whole history to be
applied. Sharing results helps in a situation where multiple parallel
fibers try to setup linearizable barrier.

Signed-off-by: Michal Maslanka <michal@redpanda.com>