rpk: dashboard: Use $__rate_interval instead of hardcoded 2m
Properly use the `$__rate_interval` when calculating rates instead of a
hardcoded `2m`.

The later breaks as soon as the prometheus scrape interval goes above
`1m`. This isn't unlikely on clusters with large partition counts.

We are already doing this for the "public metrics" dashboard.

Also increase the auto refresh interval. Values like `10s` don't make
much sense as the prometheus scrape interval is likely never that low
and just causes load on the prometheus server.