rptest: AdminOps swap kcl.DeleteRecords for rpk
- This change relies on rpk to issue delete records instead of kcl
because the rpk ducktape client harness will raise exceptions in the
case any command does not exit 0, which will cause the
AminOperationsFuzzer retry logic to kick in. This was causing some CI
tests to fail, as the retry logic would never kick-in in tests like the
PartitionBalancerTests where it is highly likely where a partition will
change leadership during the test.

- Fixes: #11950
- Fixes: #11942