cloud_storage: cache the computed result for start_offset_delta
`kafka::replicated_partition::start_offset` is called often on the fetch
path. And when cloud storage is enabled for a partition this leads to a
find operation on `segment_meta_cstore` for every call.

The result of this find doesn't change unless `_start_offset` in
`cloud_storage::partition_manifest` has changed. So this commit caches
the result of the find and updates it whenever `_start_offset` changes.