tests: more strict fix for last_applied_offset check
Wait for all victim node records to be applied. If a victim node
contains some of the records that were not applied or about to be
truncated the test should wait before selecting segments to trim as in
the case if segment contains only dirty records removing it will not
cause inconsistency.

Fixes: #8217

Signed-off-by: Michal Maslanka <michal@redpanda.com>