rptest: Correct delete-records segmt search logic
- Previously the delete_records_segment_deletion test would assert the
correct number of segmenets was deleted by comparing the low watermark
with the observed base_offsets of segments on disk.

- This is incorrect because the low watermark is a kafka offset and the
base segment offsets are redpanda offsets.

- This change fixes the issue by obtaining the translated redpanda
offset by querying the redpanda offset translator, (this having been
exposed in a previous commit) and using this offset when comparing
against base offsets of on-disk segments.