archival: Enable/disable feature by config
New parameter 'cloud_storage_disable_upload_consistency_checks' is added
to the redpanda config. The default value is 'false'. If it's set to
'true' redpanda will stop checking uploads to S3 for consistency
(default behaviour in v23.1 and earlier). When it's set to 'false'
redpanda checks every upload and generates error when the metadata of
the uploaded segment doesn't match with the actual segment or when the
metadata of the newly uploaded segment doesn't match with the previoius
segment.

This will prevent partition from being uploaded if the metadata is
seriously messed up and we have a gap in the cloud log that can't be
fixed with local data. If for some strange reason we will want to
proceed with uploads we can disable the checks.