storage: log error instead of assert when setting cloud gc offset
Causing a hard assert is not a problem per se--the method shouldn't be
invoked on non-cloud topics. However, since the assert is based on
dynamic configs where the caller invoking this method is also dependent
on dynamic configs, then there is presumably an interleaving that might
lead to this assertion firing. Ignoring with an log message is safer.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>