cloud_storage: check that STM manifest doesn't spillover during fetch
Along the fetch path, we 1) select a manifest to iterate through based on
the fetch offset, and 2) iterate through the selected manifest
starting the fetch offset.

If spillover happens in betwen 1 and 2, the fetch may end up missing
records, since at 2, scanning from below the start of the given manifest
will begin the fetch from the manifest's start.

This patch addresses this by checking after 1 and before 2 that the
STM manifest hasn't spilled over, and if so, it tries again.

Fixes #11866