cloud_storage/ducktape: Ensure segments are compactible
In stress test when asserting that some number of compacted segments
have been uploaded, this change sets a fixed cardinality for keys so
that the segments are always compactible.

This is necessary because during compacted segment reupload we do not
reupload unless the segment size has decreased.

This check can end up in preventing upload of compacted segments to the
cloud, until all uploaded compacted segments fall off the end of the
manifest and we are left with no compacted segments in the manifest,
failing the test assertions.