config: add cloud_storage_max_partition_readers_per_shard
This enables configuring the limit on partition readers
independently of the limit on segment readers.  Controlling
partition readers enables limiting tiered storage
read concurrency (e.g. because a drive is bottlenecked) without
limiting the number of segment readers we can cache (the limit
on segment readers is driven by memory)