tests/ff_test: disable leader balancer in follower fetching test
In the follower fetching tests we check if all fetch requests are
handled by the replica in the same rack as the consumer. When leadership
changes it may happen that the follower will have to contact the leader
and refresh its metadata. In this case the test would fail as there
would be some bytes sent to the follower from the node other than the
closest replica. To prevent this from happening and make the test stable
disabled the leader balancer to prevent leadership from being
transferred.

Fixes: #11134

Signed-off-by: Michal Maslanka <michal@redpanda.com>