cloud_storage: optimise data availability check
`remote_partition::is_data_available` is on the hot path of fetch
requests and is called for all of them. Previously, it called in to
`segment_meta_cstore::{size,find}`. Both are rather expensive
operations. The query can be answered without touching the cstore by
using these two observations:
1. If the start offset for the STM region is not set, the the cloud is
empty. This is the case if no segments were uploaded yet, or if all
segments were truncated by retention.
2. If the start offset is set to a value greater than the end of the
cloud log, then the cloud log is emtpy. This is the case if cloud
retention advances the start offset to truncate all segments, but GC and
the subsequent truncation did not run yet.