rptest: Make segments with no datafile in delete..
- ... records test not to be considered in final pass/fail condition

- This delete-records dt test asserts that all on disk segments have
been removed. In the case a segment exists with no data file this method
was returning 0 making the test think that a segment with a base index
of 0 still exists on disk.

- The solution is to not include these segments without a data file in
the final pass or fail condition of the test and to filter those
segments that do not have a data file out of the result set.