storage: roll up local retention reclaim estimate into usage report
Roll up the calculation of the amount of reclaimable data above local
retention threshold and make this available to callers of usage
information.

Close reviewers will notice that this is effectively a duplicate
calculation. It's true, but the places that usage information is collect
is where we want to make the reclaimable space information available
into the health monitor. The other location where this information is
calculated is in eviction schedule, which doesn't happen at a frequency
that makes sense for reporting health.

The reason these usages are not unified is merely a matter of time
constraints and managing complexity of the refactoring necessary. Once
the remainder of the features are wired up and ready we'll begin
removing this redundancy.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>
(cherry picked from commit b7785e797eb5c83836966d2a481085db9ecd117a)