cluster: report data log usage information in health report
Reports the target size, current size, and reclaimable size of the data
log storage in the health report. This is going to be used to inform the
partition balancer about how much data is reclaimable on a node, used in
cases where the disk may look full, but has data that can be removed.

When no data log storage target is set, the value returned for the
target is zero and the space manager is considered to not run.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>
(cherry picked from commit 3f307d186c23b13daff31f24a584448126994a3b)