c/balancer: force refresh health report when partition size is missing
In the case when partition is created right before the node is being
decommissioned it may be the case that its size is not present in the
health report. In this case a balancer would have to wait for at most
two metadata refresh periods to get the partition size information.
During that period the balancer backend may enter stalled mode.

Added a mechanism allowing partition balancer backend to detect
situation in which partition size is missing and force refresh the
health report. In this case a balancer tick is scheduled immediately.

Fixes: #11999

Signed-off-by: Michal Maslanka <michal@redpanda.com>