r/append_entries_buffer: use chunked fifo to store requests
Append entries buffer is used to coalesce multiple append entry requests
into single batch and reduce number of flushes. Usually the size of the
batch is small. In some case when it grows larger it may end up in
causing large contiguous allocation.

Replaced vector used to buffer pending requests with a `chunked_fifo`.
The chunk size was set to 32 as usually the size of the queue is small.

Fixes: #12072

Signed-off-by: Michal Maslanka <michal@redpanda.com>