cloud_storage: fix abort handling of remote ntp
`remote_partition` can in theory receive an abort at any time. The most
common case would be a client disconnect. Previously, this was handled
by returning the reader to the `materialized_segment_state`. This is
dangerous since the abort may catch us at a bad time and return a reader
in an undefined state. We've seen this cause issues when the reader gets
reused.