rm_stm: use system time for producer expiration tracking
Using batch timestamps for pid expiration may result in unexpected
behavior when the batches are marked with past/future timestamps.
This patch switches to using broker's system clock for the same.

Clock drift among brokers is particularly not a concern in this code
path because the default expiration timeout is 7d.