rmgmt: add feature flag for new retention space management
Even though this is feature that really only operates locally to a node
(and doesn't have the typical characteristics of a feature that requires
gating), integration with the partition balancer will benefit from this.
for example, if the cluster is in a state where an old balancer is for
an extended period of time attempt to manager new brokers that have new
local space management behavior, there may be unintended consequences as
the new brokers are not able to yet communicate to the balancer their
effective free space.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>