ssx: Add 'task_local' and 'task_local_ptr'
In many cases we can have state which is only valid before the
scheduling point. There is no way to prevent such value from being
accessed after the scheduling point. And if some method returns such
value we don't have a good way to communicate this.

The 'task_local' is a container for a single value that gets invalidated
after the current time-slice is over. The value is not destroyed until
the d-tor is called but it can't be accessed.

The 'task_local_ptr' is a non-owning smart pointer that can be used only
in the same time-slice that it was created in. If some API needs to
return a reference to the object that can be invalidated after a
scheduling point it can return the 'task_local_ptr' instead. The
'task_local_ptr' will throw an exception if used in a different
from the different reactor task (different time-slice).

Additionally, the 'task_local_ptr' implements 'value' and 'has_value'
methods. They're working similarly to std::option. The 'task_local_ptr'
can be default constructed. In this case it's equal to invalidated
'task_local_ptr'.