cloud_storage: Use 'task_local_ptr' in 'async_manifest_view'
The 'async_manifest_view_cursor' has 'manifest' method that returns a
reference to the current manifest. It's not safe to store this value in
the temporary variable and dereference it after the scheduling point
because the underlying manifest may be altered. To prevent this the
method now returns 'task_local_ptr' pointer that references the
manifest.