txns: fix list txns failure when it's used before find_coordinator
find_coordinator creates txn coordinator's topic when it's used for
the first time. list txns api expects the topic to be already created
so it used to fail when the api was invoked before find_coordinator.
Fixed the problem by creating the topic on the first list txns call.

fixes https://github.com/redpanda-data/redpanda/issues/11947