k/usage: Fix missing stats for the open bucket
- The open bucket was never correctly reporting usage statistics. It
also wasn't keeping any data in the open bucket, only filling
the value on window close, making the disk write timer useless.

- To remedy this the open bucket will be peridiocially updated when
it is desired to persist to disk and on invocations of querying the
data. That way on restart the open bucket has the expected value it
had on the previous call to persist_to_disk.