cloud_storage: compute and cache max bytes
When the max bytes configuration changes we cache the new value and
access the cached value rather than the configuration value directly
from the binding.

This change is made because in the next commit the calculation will be
expanded to include additional parameters to the computation. Without
this we'd have to recompute the value on each access, which also makes
it more awkward to decide when to log information about changes.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>
(cherry picked from commit dec333ac5fec67778421f4f2859c12503af72ff8)