config: add percent-based cloud cache size option
Introduces a percent-based expression of cloud storage max size:

   cloud_storage_cache_size_percent

along side the version expressed in number of bytes:

   cloud_storage_cache_size

Additionally, this commit configures the upgrade behavior of the
configuration options. On upgrade the defaults are:

   cloud_storage_cache_size           20_GiB
   cloud_storage_cache_size_percent   0%/nullopt

And on new systems the default are:

   cloud_storage_cache_size           0
   cloud_storage_cache_size_percent   20%

These changes mean that on upgraded systems no change in behavior or
settings are made. But on new systems, a percentage based expression is
the default.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>
(cherry picked from commit ba64314b964472fba7f17b9fa4a36e6594bc0c6b)