archival: skip attempt to advance archive start
`async_manifest_view::size_based_retention` may return an offset which
is below the start offset of the archive. In that case it doesn't make
sense to replicate the command as the STM will simply ignore it. This
patch adjusts the handling of the returned offset in `ntp_archiver` to
behave as described.

(cherry picked from commit f0dbc98605e89cc5f5bd1a9bf8df9ee9a60ad42d)