archival: skip attempt to advance archive start
`async_manifest_view::size_based_retention` may return an offset which
is below the start offset of the archive. In that case it doesn't make
sense to replicate the command as the STM will simply ignore it. This
patch adjusts the handling of the returned offset in `ntp_archiver` to
behave as described.