c/rm_stm: fixed returning incorrect `max_collectible_offset`
An off by one error in `rm_stm::max_collectible_offset` may lead to
situation in which an stm couldn't continue. If `max_collectible_offset`
returned last offset in the segment it may lead to the situation in
which the whole segment was evicted before its last entry was applied.
As the `max_collectible_offset` was returning `last_stable_offset` which
indicates the offset for which all _previous_ offsets are decided (has
no pending transactions and have been committed). In rare cases it may
lead to situation where a segment ending at offset `n` was removed even
though only `n - 1` was decided.

Signed-off-by: Michal Maslanka <michal@redpanda.com>