cloud_storage: use clean offset in archive retention
Previously, size based retention in the archive would always start
collecting from the first segment in the same spillover manifest as the
archive start offset. Since spillover manifests are not reuploaded, said
segment could have been below the archive clean offset. In this case,
retention wouldn't collect enough data to correctly satisfy the
retention policy as it's selecting the same segments as in previous
iterations.

This patch updates size based retention to skip segments below the clean
offsets when looking for the truncation point. It might look strange
that we're using the clean offset here. Intuitively, retention should
operate from the archive start offset. The intuition is correct, but the
archive size stored in the manifest is updated when the clean offset is
updated (not the archive start offset).