tests: sort nodes before stopping them
Added sorting the nodes by id before selecting the ones to stop in
`controller_availability_test` this way the test is deterministic as it
always stop the two nodes with smallest ids. This way the test stops the
current controller and at the same time guarantees that the next
candidate is the one with the highest priority from the nodes that were
left running.

Fixes: #12275

Signed-off-by: Michal Maslanka <michal@redpanda.com>