cloud_storage: Add 'sync_manifest' method to cursor
The async_manifest_view_cursor can be invalidated if it points to STM
manifest and spillover happens. In this case if the reader is going to
fetch all segments's metadata some of the segments will be skipped. To
avoid this issue the 'sync_manifest' method was added to the cursor. The
method checks if the cursor points to STM manifest and if it does and if
the start offset of the manifest differs from what's expected it fetches
the spillover manifest.

Consider the following initial state:

  [spill 0][spill 1][spill 3][           STM  ]
                              ^
                              |
                    cursor----+

Then, after spillover:

  [spill 0][spill 1][spill 3][spill 4][  STM  ]
                                       ^
                                       |
                    cursor-------------+

The cursor stores the reference to the STM manifest, but the manifest
misses offset range covered by the "spill 4" manifest which was uploaded
to S3. When the 'sync_manifest' will be called the situation will be
fixed by repositioning cursor to the "spill 4" like this:

  [spill 0][spill 1][spill 3][spill 4][  STM  ]
                              ^
                              |
                    cursor----+

The implementation caches the start offset of the STM manifest and in
the 'sync_manifest' it checks it and calls 'seek' if it doesn't match
the referenced STM manifest.

(cherry picked from commit 187af2405749e9ff9554bf7e5b1b16230fa37dc5)