archival: Handle end offset inside batch on reupload
When creating a re-upload candidate where the end offset lies inside a
batch, we adjust the upload so that the metadata indicates that the
upload ended at the end offset, but the underlying segment file contains
the full batch which contains the end offset. This means that the
segment ends after the end offset, which is resolved in the read path as
duplicate records.