storage: Allow search for end offset inside batch
The utilities to search for end offset in a segment are extended to
allow the end offset to be inside a batch, and accept the entire batch
as part of a result.

Before this, if the searched for offset was inside a batch the result
would contain the previous batch.