schema_registry/avro: Remove redundant namespaces
This PR keeps track of the current namespace in a stack, starting with
the implicitly null "empty" namespace.

If the namespace changes, push it to the stack.

If a namespace is redundant (the same as outer scope), remove it.

Fixes #11912

Signed-off-by: Ben Pope <ben@redpanda.com>
(cherry picked from commit 9cfd7bd73731c2f865e59ac91460a9608f9054c8)