cloud_storage: avoid writing empty tx ranges to cache
Previously the check for metadata_size_hint==0 was
initializing `_tx_range`, but that didn't stop `do_hydrate_txrange`
from running, attempting to download the transaction manifest,
and usually generating a 404 response and putting an empty
manifest into the cache.

After this change, segments with no tx manifest entries simply
issue no request for the tx manifest, and write nothing to
the cache for tx manifest.