storage: Do not hardcode set of batch types in storage
Currently, Raft and Tiered-storage are using the same set of batch types
for offset translation. It's encoded in the
raft::offset_translator_batch_types function. But the comapction uses
its own function called 'is_compactible' which uses hardcoded set of
types. This commit fixes this by using the same set of types everywhere.