cloud_storage_e2e_fuzz_test: initial structure
This adds the initial structure for a fuzz test, introducing some test
ops and a method to generate a sequence of ops.

The generated sequence isn't completely random: it instead prunes out
some sequences that are likely to no-op (e.g. only roll the log if there
are records expected in the active segment).

This patch only generates and logs the sequence; a later patch will
implement and run them.