cloud_storage: Handle unexpected EOF
Explicitly store the flag that indicates that the EOF was reached before
the expected last batch. Use this flag to signal EOF condition to the
top level reader to avoid infinite loop. Do not throw stuck reader
exception which makes impossible for the reader to consume the entire
partition if the missing batch is at the end of the segment (this may
happen if the segment is compacted).

(cherry picked from commit 481d4edd7358025b622d22a63385ee89393c7cd5)