rptest: increase failure injection test timeout
`StorageFailureInjectionTest.test_storage_failure_injection` failed
reliably on CDT nodes while waiting for the node to be restarted after
a storage failure injection event. Upon further investigation it turned
out that a large amount of time could be spent in the kernel in order to
clean up resources associate with the Redpanda process. During this
time, the process would be in the "uninterruptible sleep" state.

This patch makes the timeout more generous.

Fixes #12396