tests: add rpk plugin and byoc tests
This commit introduces a new go binary into the
ducktape docker image. This binary it's a mocked
plugin which outputs the passed args.

We also include a test that uses this binary to
assert that rpk is passing the required flag
to the managed plugin (byoc).

(cherry picked from commit 7a4ff7c9a8311fe462fe7c49a87f65623385470c)