c/log_eviction_stm: refactored log eviction stm to use queue
Log eviction STM previously was using a condition variable and raft
offset monitor to synchronize processing of two different eviction
events types i.e. storage driven eviction and delete record related
eviction. The coordination between the fiber executing eviction and the
the one requesting it was complicated. Changed the coordination logic to
use a queue where all the eviction requests are enqueued and the fibre
simply executes them in sequence.

Signed-off-by: Michal Maslanka <michal@redpanda.com>