tests: bump timeout for waiting on truncation
If the eviction truncation loop truncates up to the max collectible
offset but knows it has more to truncate, it will wait for 5s. Some
tests wait for 3s and expect to be truncated down to one segment. This
patch bumps these timeouts.

Alternatively, we could reduce the sleep with the expectation that max
collectible offset will progress quickly. This isn't always the case
(e.g. slow uploads, open transactions, etc).