storage: fix race between segment.ms and appends
We have seen a couple of races between the application of `segment.ms`
and the normal append path. They had the following pattern in common:
1. application of `segment.ms` begins
2. a call to `segment::append` is interleaved
3. the append finishes first and and advances the dirty offsets, which
the rolling logic in `segment.ms` does not expect
-- or --
4. `segment.ms` releases the current appender while the append is
ongoing, which the append logic does not expect

The issue is solved by introducing a mutex and grabbing it on the append
path. The same mutex is held during the application of `segment.ms` in
order to block writes.

The overhead of this solution should be minimal for two reasons:
* One (previously undocumented) invariant of the storage layer is that
there may be only one append in flight to a partition (flushes may be
concurrent however). This invariant is also respected by Raft; see
`append_entries_buffer::do_flush` where the call to
`consensus::do_append_entries` is done under the Raft op lock.
* Contention is still possible if we decide to apply `segment.ms` and
receive and wish to append while holding the lock. However, `segment.ms`
should almost never be applied in high throughput scenarios as the
segments will roll naturally in that case.